Completed the second bonus step of assignment 3
Variables where the type is set to:

    IContent | undefined

are the same as if the variable was optional. It's allowed to be an IContent value, or undefined (that's what the | symbol does in this case).

The use of === checks to make sure the variable value AND type are the same, for extra careful code comparisons