Updated the ContentList to show a loading content message
If the server takes 3 seconds to respond with the data, the user should be shown that content is waiting to load, not that there is no data.

Since undefined is a falsy value, we can make our contentList's IContent[] optional and have an initial value of undefined. This lets us check if it's falsy in the html file, display a loading message when it is undefined, and display a message that the array is empty if it loaded from the server with an empty array (since an empty array is truthy in javascript)